+++ src/Icu/MessageFormat/Parser/Type/Location.php
@@ -24,6 +24,8 @@ namespace FormatPHP\Icu\MessageFormat\Parser\Type;
 
 class Location
 {
+    use DeepCloner;
+
     public LocationDetails $start;
     public LocationDetails $end;