+++ tests/Icu/MessageFormat/Parser/Type/ArgumentElementTest.php
@@ -12,7 +12,7 @@ use FormatPHP\Test\TestCase;
 
 class ArgumentElementTest extends TestCase
 {
-    public function testType(): void
+    public function testConstructor(): void
     {
         $start = new LocationDetails(0, 1, 1);
         $end = new LocationDetails(2, 4, 6);
@@ -21,5 +21,21 @@ class ArgumentElementTest extends TestCase
         $element = new ArgumentElement('argument value', $location);
 
         $this->assertEquals(ElementType::Argument(), $element->type);
+        $this->assertSame('argument value', $element->value);
+        $this->assertSame($location, $element->location);
+    }
+
+    public function testDeepClone(): void
+    {
+        $start = new LocationDetails(0, 1, 1);
+        $end = new LocationDetails(2, 4, 6);
+        $location = new Location($start, $end);
+
+        $element = new ArgumentElement('argument value', $location);
+        $clone = clone $element;
+
+        $this->assertNotSame($location, $clone->location);
+        $this->assertNotSame($start, $clone->location->start);
+        $this->assertNotSame($end, $clone->location->end);
     }
 }